Fix MacOS build

+++ examples/server/httplib.h
@@ -3495,7 +3495,7 @@ inline bool read_content_with_length(Stream &strm, uint64_t len,
 
   uint64_t r = 0;
   while (r < len) {
-    auto read_len = len - r;
+    auto read_len = static_cast<std::size_t>(len - r);
     auto n = strm.read(buf, (std::min)(read_len, CPPHTTPLIB_RECV_BUFSIZ));
     if (n <= 0) { return false; }
